requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.12.3
blinker==1.9.0
blueprint==3.4.2
bs4==0.0.2
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==44.0.0
curl_cffi==0.11.1
cycler==0.12.1
DateTime==5.5
Deprecated==1.2.15
filelock==3.18.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
frozendict==2.4.6
fsspec==2025.5.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
hf-xet==1.1.2
holidays==0.63
httplib2==0.22.0
huggingface-hub==0.32.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jpype1==1.5.2
kiwisolver==1.4.7
konlpy==0.6.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
msgpack==1.1.0
multipledispatch==1.0.0
multitasking==0.0.11
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==24.2
pandas==2.2.3
peewee==3.18.1
pillow==11.0.0
platformdirs==4.3.8
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
PyKakao==0.0.7
pykrx==1.0.48
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
six==1.17.0
soupsieve==2.6
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
typing_extensions==4.13.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.0
xlrd==2.0.1
yfinance==0.2.61
zope.interface==7.2
